ncsi: Don't report an error when clearing the initial state of nonexistent channels.
The NCSI specification implies that when a channel does not exist, no response is given.
This behavior is consistent with the proprietary firmware.

When a response is given, the linux kernel spews (~30 times) "NCSI: 'bad' packet
ignored for type 0x80" on an invalid response. Disable the response to silence
these messages and make the behavior match the proprietary firmware.