README.md: rework usage and development sections
The "Usage" section now reflects the structure of the release archive and is moved earlier in the page to be aimed at new users, while still pointing to later sections where necessary. The flash commands use the `-i eth` command to avoid the extra step of unbinding and losing connectivity on a port.

"Development" now includes more detailed instructions on unbinding the driver and the implications thereof.